Create Ceil the floor
Given an unsorted array a[] of nintegers and an integer x, find floor and ceiling of X in a[0..N-1].

Floor of x is the largest element which is smaller than or equal to x. Floor of x doesn’t exist if x is smaller than smallest element of a[].

Ceil of x is the smallest element which is greater than or equal to x. Ceil of x doesn’t exist if x is greater than greatest element of a[].