Create Equilibrium_Point
Given an array A of n positive numbers. The task is to find the first Equilibrium Point in an array. 
Equilibrium Point in an array is a position such that the sum of elements before it is equal to the sum of elements after it.

Question link--https://practice.geeksforgeeks.org/problems/equilibrium-point-1587115620/1?page=1&category[]=Arrays&sortBy=submissions